feat: 订阅计划、OAuth刷新、IP地理、每日配额等新功能，全栈大量改进
新增后端功能：
- SubscriptionPlan 控制器/服务/仓库
- UserQuota 控制器
- DailyQuotaService / DailyQuotaUsageRepository
- OAuthTokenRefreshService
- IpGeoService
- TempDisabledProbeService

前端新增页面：MySubscription、SubscriptionPlans

全面更新：后端模型、服务、仓库、控制器；前端所有页面组件；
前端构建产物更新到 static/assets；新增测试和文档。

+++ backend/src/main/java/com/firstapi/backend/model/RecordsData.java
@@ -39,9 +39,9 @@ public class RecordsData {
 
     public static class BarPoint {
         public String name;
-        public int tokens;
+        public double tokens;
 
-        public BarPoint(String name, int tokens) {
+        public BarPoint(String name, double tokens) {
             this.name = name;
             this.tokens = tokens;
         }
